Removing Cyclic Redundancy
