+++ retriever/lib/engine.py
@@ -18,7 +18,7 @@ import csv
 import re
 import time
 from urllib.request import urlretrieve
-from retriever import open_fr, open_fw, open_csvw
+from retriever.lib.scripts import open_fr, open_fw, open_csvw
 from retriever.lib.defaults import DATA_SEARCH_PATHS, DATA_WRITE_PATH
 from retriever.lib.cleanup import no_cleanup
 from retriever.lib.warning import Warning
