+++ scripts/breed_bird_survey_50stop.py
@@ -15,7 +15,7 @@ import zipfile
 from decimal import Decimal
 from retriever.lib.templates import Script
 from retriever.lib.models import Table, Cleanup, correct_invalid_value
-from retriever import open_fr, open_fw
+from retriever.lib.scripts import open_fr, open_fw
 from retriever.lib.defaults import VERSION
 from pkg_resources import parse_version
