+++ scripts/forest_inventory_analysis.py
@@ -11,7 +11,7 @@ import os
 from retriever.lib.templates import Script
 from retriever.lib.models import Table
 from retriever.lib.defaults import VERSION
-from retriever import open_fr, open_fw
+from retriever.lib.scripts import open_fr, open_fw
 from pkg_resources import parse_version
