+++ scripts/npn.py
@@ -15,7 +15,7 @@ from retriever.lib.models import Table
 from pkg_resources import parse_version
 import xml.etree.ElementTree as ET
 import datetime
-from retriever import open_fw, open_csvw
+from retriever.lib.scripts import open_fw, open_csvw
 from retreiver.lib.defualts import VERSION, DATA_WRITE_PATH
