+++ scripts/plant_life_hist_eu.py
@@ -3,7 +3,7 @@
 from builtins import str
 from retriever.lib.models import Table, Cleanup, correct_invalid_value
 from retriever.lib.templates import Script
-from retriever import open_fr, open_fw
+from retriever.lib.scripts import open_fr, open_fw
 from retriever.lib.defaults import VERSION
 from pkg_resources import parse_version
